+++ pkg/forwarder/blocked_endpoints.go
@@ -6,16 +6,21 @@
 package forwarder
 
 import (
+	"math"
+	"math/rand"
 	"sync"
 	"time"
+
+	log "github.com/cihub/seelog"
+
+	"github.com/DataDog/datadog-agent/pkg/config"
 )
 
-// Whether or not our config has been loaded by the agent yet.
-var configLoaded = false
+const secondsFloat = float64(time.Second)
 
-// This is the number of errors it will take to reach the maxBackoffTime. Our
-// blockedEndpoints circuit breaker uses this value as the maximum number of errors.
-var maxErrors int
+func randomBetween(min, max float64) float64 {
+	return rand.Float64()*(max-min) + min
+}
 
 type block struct {
 	nbError int
@@ -25,17 +30,69 @@ type block struct {
 type blockedEndpoints struct {
 	errorPerEndpoint map[string]*block
 	m                sync.RWMutex
+
+	// This controls the overlap between consecutive retry interval ranges. When
+	// set to `2`, there is a guarantee that there will be no overlap. The overlap
+	// will asymptotically approach 50% the higher the value is set.
+	minBackoffFactor float64
+
+	// This controls the rate of exponential growth. Also, you can calculate the start
+	// of the very first retry interval range by evaluating the following expression:
+	// baseBackoffTime / minBackoffFactor * 2
+	baseBackoffTime float64
+
+	// This is the maximum number of seconds to wait for a retry.
+	maxBackoffTime float64
+
+	// This controls how many retry interval ranges to step down for an endpoint
+	// upon success. Increasing this should only be considered when maxBackoffTime
+	// is particularly high or if our intake team is particularly confident.
+	recoveryInterval int
+
+	// This derived value is the number of errors it will take to reach the maxBackoffTime.
+	maxErrors int
 }
 
 func newBlockedEndpoints() *blockedEndpoints {
-	// All forwarder settings are used directly or indirectly (GetBackoffDuration) by
-	// this circuit breaker singleton. Therefore, it makes sense to load them here.
-	if !configLoaded {
-		loadConfig()
-		configLoaded = true
+	backoffFactor := config.Datadog.GetFloat64("forwarder_backoff_factor")
+	if backoffFactor < 2 {
+		log.Warnf("Configured forwarder_backoff_factor (%v) is less than 2; 2 will be used", backoffFactor)
+		backoffFactor = 2
 	}
 
-	return &blockedEndpoints{errorPerEndpoint: make(map[string]*block)}
+	backoffBase := config.Datadog.GetFloat64("forwarder_backoff_base")
+	if backoffBase <= 0 {
+		log.Warnf("Configured forwarder_backoff_base (%v) is not positive; 2 will be used", backoffBase)
+		backoffBase = 2
+	}
+
+	backoffMax := config.Datadog.GetFloat64("forwarder_backoff_max")
+	if backoffMax <= 0 {
+		log.Warnf("Configured forwarder_backoff_max (%v) is not positive; 64 seconds will be used", backoffMax)
+		backoffMax = 64
+	}
+
+	errorsMax := int(math.Floor(math.Log2(backoffMax/backoffBase))) + 1
+
+	recInterval := config.Datadog.GetInt("forwarder_recovery_interval")
+	if recInterval <= 0 {
+		log.Warnf("Configured forwarder_recovery_interval (%v) is not positive; 1 will be used", recInterval)
+		recInterval = 1
+	}
+
+	recoveryReset := config.Datadog.GetBool("forwarder_recovery_reset")
+	if recoveryReset {
+		recInterval = errorsMax
+	}
+
+	return &blockedEndpoints{
+		errorPerEndpoint: make(map[string]*block),
+		minBackoffFactor: backoffFactor,
+		baseBackoffTime:  backoffBase,
+		maxBackoffTime:   backoffMax,
+		recoveryInterval: recInterval,
+		maxErrors:        errorsMax,
+	}
 }
 
 func (e *blockedEndpoints) close(endpoint string) {
@@ -50,10 +107,10 @@ func (e *blockedEndpoints) close(endpoint string) {
 	}
 
 	b.nbError++
-	if b.nbError > maxErrors {
-		b.nbError = maxErrors
+	if b.nbError > e.maxErrors {
+		b.nbError = e.maxErrors
 	}
-	b.until = time.Now().Add(GetBackoffDuration(b.nbError))
+	b.until = time.Now().Add(e.getBackoffDuration(b.nbError))
 
 	e.errorPerEndpoint[endpoint] = b
 }
@@ -69,11 +126,11 @@ func (e *blockedEndpoints) recover(endpoint string) {
 		b = &block{}
 	}
 
-	b.nbError -= recoveryInterval
+	b.nbError -= e.recoveryInterval
 	if b.nbError < 0 {
 		b.nbError = 0
 	}
-	b.until = time.Now().Add(GetBackoffDuration(b.nbError))
+	b.until = time.Now().Add(e.getBackoffDuration(b.nbError))
 
 	e.errorPerEndpoint[endpoint] = b
 }
@@ -86,4 +143,22 @@ func (e *blockedEndpoints) isBlock(endpoint string) bool {
 		return true
 	}
 	return false
+}
+
+func (e *blockedEndpoints) getBackoffDuration(numErrors int) time.Duration {
+	var backoffTime float64
+
+	if numErrors > 0 {
+		backoffTime = e.baseBackoffTime * math.Pow(2, float64(numErrors))
+
+		if backoffTime > e.maxBackoffTime {
+			backoffTime = e.maxBackoffTime
+		} else {
+			min := backoffTime / e.minBackoffFactor
+			max := math.Min(e.maxBackoffTime, backoffTime)
+			backoffTime = randomBetween(min, max)
+		}
+	}
+
+	return time.Duration(backoffTime * secondsFloat)
 }
